+++ src/colormodes/dark.js
@@ -1,0 +1,8 @@
+var MS = require('../ms.js');
+module.exports = MS.colorMode(
+  'rgb(80,0,80)',
+  'rgb(0,107,140)',
+  'rgb(200,0,0)',
+  'rgb(0,160,0)',
+  'rgb(225,220,0)'
+);