+++ src/colormodes/framecolor.js
@@ -1,0 +1,8 @@
+var MS = require('../ms.js');
+module.exports = MS.colorMode(
+  'rgb(255,0,255)',
+  'rgb(0, 255, 255)',
+  'rgb(255, 0, 0)',
+  'rgb(0, 255, 0)',
+  'rgb(255, 255, 0)'
+);