+++ src/colormodes/light.js
@@ -1,0 +1,8 @@
+var MS = require('../ms.js');
+module.exports = MS.colorMode(
+  'rgb(255,161,255)',
+  'rgb(128,224,255)',
+  'rgb(255,128,128)',
+  'rgb(170,255,170)',
+  'rgb(255,255,128)'
+);