+++ src/colormodes/medium.js
@@ -1,0 +1,8 @@
+var MS = require('../ms.js');
+module.exports = MS.colorMode(
+  'rgb(128,0,128)',
+  'rgb(0,168,220)',
+  'rgb(255,48,49)',
+  'rgb(0,226,110)',
+  'rgb(255,255,0)'
+);