+++ src/colormodes/offwhite.js
@@ -1,0 +1,8 @@
+var MS = require('../ms.js');
+module.exports = MS.colorMode(
+  'rgb(239, 239, 239)',
+  'rgb(239, 239, 239)',
+  'rgb(239, 239, 239)',
+  'rgb(239, 239, 239)',
+  'rgb(239, 239, 239)'
+);